Delete confirms added

+++ src/main/webapp/organizationInfo.jsp
@@ -21,11 +21,20 @@
    }
    %>
 <%@include file = "layout1.jsp"%>
+<script>
+    function confirmDelete() {
+        if (confirm("Вы уверены, что хотите удалить организацию?")) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+</script>
 <br>
     <div class="form-block center-block" style="width: 50%; min-height: 700px;">
         <div style="margin-left: 90%;">
             <p style="font-size: 25px;"><a href="changeOrganizationInfo.jsp?id=<%=org.getId()%>"><i class="fa fa-edit"></i></a>
-            <a href="deleteOrganization.jsp?id=<%=org.getId()%>"><i class="fa fa-close"></i></a></p>
+            <a href="deleteOrganization.jsp?id=<%=org.getId()%>" onclick="return confirmDelete();"><i class="fa fa-close"></i></a></p>
         </div>
         <center><h2 class="title"><%=org.getName()%></h2></center>
         <hr>
